feat: Refactor code to ensure data.db is always an Array
Previously, an if statement checking for data.db's existence was followed by the execution of a data.db assignment conditional. Now, when there is an existing data.db, the code checks if it is an array. If it's not an array, create a new array with the non-array value as the sole element. This ensures that data.db is always an array, simplifying processing later on in the code.