feat: Supports multiple databases to be used independently or as a mesh.
Refactor `CoCreateCrudServer`'s database property to db and replace the parameter value across code for consistency. Use `crud` instead of `db` in `wsManager` event listeners and function calls. Add default database to data parameter when it's not defined.

This commit changes the name of `database` property to `db` in `CoCreateCrudServer` constructor and replaces its usage across the code for consistency. It also replaces all references to `db` in `wsManager` event listeners and function calls with `crud`. Additionally, default database name is added to `data` parameter when it's not defined.